pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mdpp"
dynamic = ["version"]
authors = [{ name = "Zhaoyang Li", email = "zhaoyangli@stanford.edu" }]
license = { file = "LICENSE" }
readme = "README.md"
description = "MD simulation pre- and post-processing"
requires-python = ">=3.12"
dependencies = [
    "ParmEd",
    "Pillow",
    "biopython",
    "biotite",
    "deeptime",
    "matplotlib",
    "mdanalysis",
    "mdtraj",
    "mplplots @ git+https://github.com/FridrichMethod/mplplots.git",
    "numba",
    "numpy",
    "pandas",
    "panedr",
    "pdb-tools",
    "pdb2pqr",
    "plotly",
    "polars",
    "prody",
    "propka",
    "rdkit",
    "scikit-learn",
    "scipy",
    "seaborn",
    "tqdm",
    "uv",
]

[project.scripts]
mdpp = "mdpp._cli:main"

[tool.setuptools.dynamic]
version = { attr = "mdpp.__version__" }

[project.optional-dependencies]
viz = ["gmxvg", "jupyter", "py3dmol", "nglview"]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-timeout",
    "pytest-xdist",
    "ruff",
]
docs = ["mkdocs-material", "mkdocstrings[python]"]
mypy = ["mypy"]
openmm = ["openmm[cuda12]"]

[tool.ruff]
target-version = "py312"
preview = true
src = ["src"]
line-length = 100
force-exclude = true
extend-exclude = [
    ".git",
    ".hg",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
]

[tool.ruff.lint]
# Check https://beta.ruff.rs/docs/rules/ for full list of rules
select = [
    "A",    # flake8-builtins
    "ARG",  # flake8-unused-arguments
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "C90",  # mccabe
    "E",    # pycodestyle
    "F",    # Pyflakes
    "I",    # isort
    "NPY",  # numpy
    "PD",   # pandas-vet
    "PERF", # perflint
    "PLE",  # pylint-error
    "RUF",  # ruff
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
    "UP",   # pyupgrade
]
extend-select = [
    "D",   # pydocstyle
    "W",   # pycodestyle
    "PLW", # pylint-warning
]
ignore = ["B905", "E203", "E501", "E741", "UP037"]
explicit-preview-rules = true

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
known-first-party = ["mdpp"]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"*.ipynb" = ["F401", "RUF059"]
"src/mdpp/data/mdps/__init__.py" = ["D104"]
"tests/**" = ["D101", "D102", "D103", "D104"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
pretty = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = false
namespace_packages = false
enable_incomplete_feature = ["Unpack"]
disable_error_code = ["import-not-found", "import-untyped", "no-any-return"]

[[tool.mypy.overrides]]
module = [
    "Bio.*",
    "IPython.*",
    "PIL.*",
    "matplotlib.*",
    "mdtraj.*",
    "numba.*",
    "numpy.*",
    "openmm.*",
    "tqdm.*",
    "pandas.*",
    "pdbfixer.*",
    "pytest.*",
    "panedr.*",
    "rdkit.*",
    "scipy.*",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.setuptools]
include-package-data = true
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"mdpp.data" = ["**/*.mdp"]

[tool.pytest.ini_options]
addopts = "-v -s --tb=short --strict-markers --color=yes -n auto"
testpaths = ["tests"]
pythonpath = ["src"]
env = ["TQDM_DISABLE=1"]
